code/src/requirements.txt
# FastAPI and web server
fastapi==0.114.0
uvicorn==0.28.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0  # HTTP client for API calls

# Database
pymongo==4.6.2
motor==3.3.2

# Authentication
PyJWT==2.10.1
passlib==1.7.4
bcrypt==4.1.2
cryptography==42.0.5
python-jose[cryptography]==3.4.0
email-validator==2.1.1

# ML and RAG
# transformers==4.41.0 - removed as not used directly
torch==2.2.1
sentence-transformers==2.6.0
# huggingface_hub==0.23.0 - removed as not used directly
pillow==10.3.0

# Data processing
numpy==1.26.4
pandas==2.2.2

# Utilities
pydantic==2.7.1
pydantic-settings==2.1.0
jinja2==3.1.4

# New dependencies
redis==5.0.1
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-core==0.1.10
chromadb==0.4.18
tiktoken>=0.5.2,<0.6.0
tenacity==8.2.3
starlette-csrf==3.0.0 